+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/remainingnumber/subhdmana/AddSubHdManagementService.java
@@ -3,7 +3,6 @@
  */
 package nts.uk.ctx.at.shared.dom.remainingnumber.subhdmana;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -81,18 +80,18 @@ public class AddSubHdManagementService {
 			}
 			if (subHdManagementData.getCheckedHoliday() && subHdManagementData.getCheckedSubHoliday()) {
 				// ドメインモデル「振休休出振付け管理」に紐付きチェックされているもの全てを追加する
-				BigDecimal usedDays = BigDecimal.valueOf(0);
+				Double usedDays = 0d;
 				int targetSelectionAtr = 2; // 固定値：手動
 				int usedHours = 0;
 				Double usedDay = subHdManagementData.getSelectedCodeOptionSubHoliday()
 						+ subHdManagementData.getSelectedCodeOptionSubHoliday();
 				if (subHdManagementData.getCheckedSplit()) {
-					usedDays = BigDecimal.valueOf(usedDay);
+					usedDays = usedDay;
 					LeaveComDayOffManagement domainLeaveComDayOffManagementSub = new LeaveComDayOffManagement(leaveId,
 							comDayOffIDSub, usedDays, usedHours, targetSelectionAtr);
 					repoLeaveComDayOffMana.add(domainLeaveComDayOffManagementSub);
 				} else {
-					usedDays = BigDecimal.valueOf(subHdManagementData.getSelectedCodeOptionSubHoliday());
+					usedDays = subHdManagementData.getSelectedCodeOptionSubHoliday();
 				}
 				LeaveComDayOffManagement domainLeaveComDayOffManagement = new LeaveComDayOffManagement(leaveId,
 						comDayOffID, usedDays, usedHours, targetSelectionAtr);
